fix compiler warnings
- Mark RunableImpl, TaskImpl, Map::Node, Helper, HeapContImpl, ExtStackContImpl as final
- Add virtual ~ContImpl() for correct polymorphic delete
- Add override to threadId() and nativeHandle()
- Fix reorder-init-list in http_srv_ut.cpp (wg before addrLen)
- Fix unused variables and parameters in various test files
- Fix sign-compare warnings with explicit casts
- Fix -Wparentheses in in_zc.cpp
- Fix -Wmacro-redefined for _GNU_SOURCE in view.cpp
- Fix -Winvalid-noreturn with __builtin_unreachable()
- Fix -Wself-assign-overloaded in view_ut.cpp
- Add alignas(16) to Bits and State in wait_queue.cpp
- Add missing .revents initializer in poller.cpp
- Remove unused lambda captures

+++ std/mem/mem_pool.cpp
@@ -83,5 +83,5 @@ void MemoryPool::allocateNewChunk(size_t minSize) {
     currentChunkEnd = (u8*)newChunk + nextChunkSize;
     currentChunk = (u8*)(newChunk + 1);
 
-    STD_ASSERT(currentChunkEnd - currentChunk >= minSize);
+    STD_ASSERT((size_t)(currentChunkEnd - currentChunk) >= minSize);
 }
